% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{connect_to_samples}
\alias{connect_to_samples}
\title{Creates an authenticated API client to an AQUARIUS Samples account}
\usage{
connect_to_samples(url, api_token)
}
\arguments{
\item{url}{Base URL for the AQUARIUS Samples account. "https://{yourinstance}.aqsamples.com".}

\item{api_token}{API token for the AQUARIUS Samples account. Browse to "https://{yourinstance}.aqsamples.com/api" to retrieve the account's api_token.}
}
\value{
An authenticated client for the AQUARIUS Samples account.
}
\description{
Builds an a dynamic, authenticated client for your AQUARIUS Samples instance.
}
\examples{
Connect to AQUARIUS Samples and retrieve names of the labs
}
