% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{content_or_stop}
\alias{content_or_stop}
\title{Stops when an HTTP error is received}
\usage{
content_or_stop(response, ...)
}
\arguments{
\item{response}{An httr response object.}

\item{...}{Optional parameters for the httr:content() function.}
}
\value{
The httr response object.
}
\description{
Returns a successful HTTP response, but stops if an error is received.
}
