% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{json_or_stop}
\alias{json_or_stop}
\title{Returns a dataframe of the JSON response}
\usage{
json_or_stop(response)
}
\arguments{
\item{response}{An httr response object.}
}
\value{
A dataframe of the JSON response object.
}
\description{
Returns a data frame of the JSON response object, but stops if an error is received.
}
